requirements.txt
streamlit==1.40.1
openai==1.54.4
python-dotenv==1.0.1
